pyproject.toml
[project]
name = 'casino'
version = '0.1.0'
description = "Big Bungo's Casino - One-Dollar Bet Tracker"
requires-python = '>=3.13'
dependencies = [
  'discord.py==2.6.4',
  'python-dotenv==1.2.1',
  'pytz==2025.2',
  "yoyo-migrations>=9.0.0",
]

[project.optional-dependencies]
dev = [
  'pytest==9.0.2',
  'pytest-asyncio>=0.26.0',
  'ruff==0.14.9',
]

[tool.ruff]
line-length = 100
target-version = 'py313'

[tool.ruff.lint]
select = ['E', 'F', 'W', 'Q']
ignore = ['E501']

[tool.ruff.lint.flake8-quotes]
inline-quotes = 'single'
multiline-quotes = 'single'

[tool.ruff.format]
quote-style = 'single'
